+++ src/lib.rs
@@ -1,11 +1,14 @@
 #![doc = include_str!("../doc/lib.md")]
 
 mod into_iter;
+mod iterators;
 mod slot;
 
-use std::{hint::unreachable_unchecked, ops::{Index, IndexMut}, mem::replace};
+use std::{hint::unreachable_unchecked, mem::replace, ops::{Index, IndexMut}};
 use into_iter::FreelistIter;
 use slot::Slot;
+
+use iterators::*;
 
 
 #[doc = include_str!("../doc/freelist.md")]
@@ -77,13 +80,28 @@ impl<T> Freelist<T> {
     /// use fffl::Freelist;
     /// 
     /// let mut fl = Freelist::from([1, 2]);
-    /// let _ = fl.push(3);
+    /// 
+    /// let _ = fl.push(3); // Pushes to the back so the returned index is trivially `2`
+    /// 
     /// assert_eq!(fl.to_vec(), vec![1, 2, 3]);
+    /// ```
+    /// Push keeps track of previously freed slots and will use those first if available:
+    /// ```
+    /// use fffl::Freelist;
+    /// 
+    /// let mut fl = Freelist::from([1, 2, 3, 4]);
+    /// 
+    /// let _ = fl.remove(1); // Some(2)
+    /// let _ = fl.remove(2); // Some(3)
+    /// 
+    /// assert_eq!(fl.push(13), 2);
+    /// assert_eq!(fl.push(14), 1);
+    /// assert_eq!(fl.to_vec(), [1, 14, 13, 4]); 
     /// ```
     /// # Time complexity
     /// 
     /// Takes amortized *O*(1) time.  The freelist will first try to push into 
-    /// previously freed slots before reallocating, if available.  *O*(*capacity*) time is taken to copy the
+    /// previously freed slots (if available) before reallocating.  *O*(*capacity*) time is taken to copy the
     /// freelist's elements to a larger allocation. This expensive operation is
     /// offset by the *capacity* *O*(1) insertions it allows.
     #[inline]
@@ -105,8 +123,18 @@ impl<T> Freelist<T> {
 
     /// Returns the next available index.
     /// 
-    /// If there are no free slots, this will be the length of the freelist and 
-    /// therefore NOT a suitable slot to index.
+    /// If there are no free slots, this will be the length of the freelist.
+    /// 
+    /// # Example
+    /// ```
+    /// use fffl::Freelist;
+    /// 
+    /// let mut fl = Freelist::from([1, 2, 3]);
+    /// 
+    /// assert_eq!(fl.next_available(), 3);
+    /// let _ = fl.remove(1); // Some(2)
+    /// assert_eq!(fl.next_available(), 1);
+    /// ```
     #[inline]
     pub fn next_available(&self) -> usize {
         match self.next {
@@ -229,6 +257,18 @@ impl<T> Freelist<T> {
     }
 
     /// Converts the freelist into a `Vec<T>`, skipping free slots.
+    /// 
+    /// # Examples
+    /// 
+    /// ```
+    /// use fffl::Freelist;
+    /// 
+    /// let mut fl = Freelist::from([1, 2, 3, 4]);
+    /// let _ = fl.remove(0); // Some(1)
+    /// let _ = fl.remove(2); // Some(3)
+    /// 
+    /// assert_eq!(fl.to_vec(), [2, 4]);
+    /// ```
     pub fn to_vec(self) -> Vec<T> { 
         self.into_iter().collect() 
     }
@@ -256,11 +296,10 @@ impl<T> Freelist<T> {
 
     /// Swaps all values to the front of the freelist.
     /// 
-    /// After repeated calls to [`push`] and [`remove`], caching and index 
+    /// After repeated calls to [`push`](Freelist::push) and [`remove`](Freelist::remove), caching and index 
     /// peformance may worsen as a result of increasingly randomized 
-    /// reads/writes to the freelist. This functions addresses this by 
-    /// shuffling values from the back of the freelist into free slots at 
-    /// the front.
+    /// reads/writes to the freelist. This is addressed by shuffling values 
+    /// from the back of the freelist into free slots at the front.
     /// 
     /// The function under the hood works thus:
     /// ```text
@@ -271,12 +310,12 @@ impl<T> Freelist<T> {
     /// Before: freelist[V, F, F, V, V, F, V]
     /// After:  freelist[V, V, V, V, F, F, F]
     /// ```
-    /// # Important Note
+    /// # Important
     /// 
-    /// This function does **NOT** retain the original order and `indices` 
-    /// returned by previous calls to [`push`] may no longer refer to their 
+    /// This will **NOT** retain order, meaning `indices` 
+    /// returned by previous calls to [`push`](Freelist::push) may no longer refer to their 
     /// values, therefore usefulness is somewhat limited.  The 
-    /// upside is repeated calls to [`push`] may now have better 
+    /// upside is repeated calls to [`push`](Freelist::push) may now have better 
     /// caching and index performance.
     /// 
     /// This is likely only beneficial for particularly sparse, large `Freelists`.
@@ -284,7 +323,7 @@ impl<T> Freelist<T> {
     /// # Time Complexity
     /// 
     /// In general this function is guaranteed to have *O*(n) performance,
-    /// where `n` is the size of the freelist. See [`size`]
+    /// where `n` is the size of the freelist. See [`size`](Freelist::size).
     pub fn compactify(&mut self) {
 
         let mut iter = self.slots.iter_mut();
@@ -404,17 +443,16 @@ impl<T> Freelist<T> {
     /// ```
     /// use fffl::Freelist;
     /// 
-    /// let mut fl = Freelist::from([1, 2, 4]);
+    /// let mut fl = Freelist::from([1, 2, 4, 8]);
+    /// let _ = fl.remove(2); // Some(4)
     /// let mut iterator = fl.iter();
     /// 
     /// assert_eq!(iterator.next(), Some(&1));
     /// assert_eq!(iterator.next(), Some(&2));
-    /// assert_eq!(iterator.next(), Some(&4));
+    /// assert_eq!(iterator.next(), Some(&8));
     /// assert_eq!(iterator.next(), None);
     /// ```
-    pub fn iter(&self) -> impl Iterator<Item = &T> {
-        self.slots.iter().filter_map(Option::from)
-    }
+    pub fn iter(&self) -> impl Iterator<Item = &T> { IterFl::new(&self.slots) }
 
     /// Returns an iterator over the full freelist that allows modifying each value.
     /// 
@@ -424,16 +462,15 @@ impl<T> Freelist<T> {
     /// ```
     /// use fffl::Freelist;
     /// 
-    /// let mut fl = Freelist::from([1, 2, 4]);
+    /// let mut fl = Freelist::from([1, 3, 5, 7]);
+    /// let _ = fl.remove(2); // Some(5)
     /// for val in fl.iter_mut() {
-    ///     *val += 2;
+    ///     *val += 1;
     /// }
     /// 
-    /// assert_eq!(fl.to_vec(), [3, 4, 6]);
+    /// assert_eq!(fl.to_vec(), [2, 4, 8]);
     /// ```
-    pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut T> {
-        self.slots.iter_mut().filter_map(Option::from)
-    }
+    pub fn iter_mut(&mut self) -> impl Iterator<Item = &mut T> { IterMutFl::new(&mut self.slots) }
 
 }
 
@@ -458,6 +495,8 @@ impl<T> Index<usize> for Freelist<T> {
         }
     }
 }
+
+
 
 impl<T> IndexMut<usize> for Freelist<T> {
 
@@ -514,7 +553,6 @@ impl<T> FromIterator<T> for Freelist<T> {
         }
     }
 }
-